+++ CHANGES.rst
@@ -93,6 +93,9 @@ Bug fixes:
 New features:
 
 - Link widget: add ``placeholder`` attributes for external and email link input fields.
+- PLIP 2150: Use ISO8601/HTML5 datetime-local format.
+  [thet]
+
 - Change the input type for date fields to ``date`` and for datetime fields to ``datetime-local``.
 - PLIP 2150: Use ISO8601/HTML5 datetime-local format.
   [thet]